// src/types/strategy.ts
// Generated by https://quicktype.io

export interface StrategySchema {
  account_id: string;
  name: string;
  symbol: string;
  symbol_id: string;
  is_future: boolean;
  leverage: number;
  entry_size: number;
  is_percentage: boolean;
  capital: number;
}

// Generated by https://quicktype.io
// Generated by https://quicktype.io

export interface Strategy {
  id: string;
  account_name: string;
  name: string;
  webhook_id: string;
  webhook_key: string;
  symbol: string;
  symbol_id: string;
  is_future: boolean;
  leverage: number;
  entry_size: number;
  is_percentage: boolean;
  capital: number;
  creation_date: string;
  invested_7_days: number;
  income_7_days: number;
  invested_7_days_percent: number;
  income_7_days_percent: number;
  nb_orders_7days: number;
}

// Generated by https://quicktype.io

export interface PublicStrat {
  id:                          string;
  name:                        string;
  webhook_id:                  string;
  webhook_key:                 string;
  symbol:                      string;
  symbol_id:                   string;
  is_future:                   boolean;
  leverage:                    number;
  capital:                     number;
  created_at:                  string;
  backtesting_start_date:      string;
  backtesting_end_date:        string;
  backtesting_initial_capital: number;
  net_profit:                  number;
  total_closed_trades:         number;
  percentage_profitable:       number;
  max_drawdown:                number;
}

export interface PublicStratPaginated {
  page_number : number;
  page_size : number;
  strategies : PublicStrat[];
  total_records : number
}